move CommonsChunkPlugin from webpack.prod.js to base.js
